pyproject.toml
[tool.poetry]
name = "llm-project"
version = "0.1.0"
description = ""
authors = ["marie0501 <97261080+marie0501@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.38.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
